backend_final/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas>=2.0.0
pillow>=10.0.0
google-genai>=0.2.0
matplotlib>=3.7.0
scikit-image>=0.21.0
numpy>=1.24.0
werkzeug>=3.0.0
Flask-SQLAlchemy>=3.1.1
pymysql>=1.1.0
